# target/generated-sources/openapi/src/main/java/com/gft/pricing/infrastructure/web/api/DefaultApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.2.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.gft.pricing.infrastructure.web.api;

import com.gft.pricing.infrastructure.web.model.generated.ApiError;
import com.gft.pricing.infrastructure.web.model.generated.PriceResponse;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-01-20T10:34:28.719132+01:00[Europe/Madrid]")
@Validated
@Tag(name = "Default", description = "the Default API")
public interface DefaultApi {

    /**
     * GET /api/v1/prices : Obtiene el precio vigente para un producto
     *
     * @param brandId Identificador de la cadena (1 &#x3D; ZARA) (required)
     * @param productId Identificador del producto (required)
     * @param applicationDate Fecha y hora en la que se desea aplicar el precio (required)
     * @return Precio encontrado (status code 200)
     *         or Petición inválida (status code 400)
     *         or No existe precio aplicable (status code 404)
     */
    @Operation(
        operationId = "findPrice",
        summary = "Obtiene el precio vigente para un producto",
        responses = {
            @ApiResponse(responseCode = "200", description = "Precio encontrado", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = PriceResponse.class))
            }),
            @ApiResponse(responseCode = "400", description = "Petición inválida", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ApiError.class))
            }),
            @ApiResponse(responseCode = "404", description = "No existe precio aplicable", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ApiError.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/api/v1/prices",
        produces = { "application/json" }
    )
    
    ResponseEntity<PriceResponse> findPrice(
        @NotNull @Min(1L) @Parameter(name = "brandId", description = "Identificador de la cadena (1 = ZARA)", required = true, in = ParameterIn.QUERY) @Valid @RequestParam(value = "brandId", required = true) Long brandId,
        @NotNull @Min(1L) @Parameter(name = "productId", description = "Identificador del producto", required = true, in = ParameterIn.QUERY) @Valid @RequestParam(value = "productId", required = true) Long productId,
        @NotNull @Parameter(name = "applicationDate", description = "Fecha y hora en la que se desea aplicar el precio", required = true, in = ParameterIn.QUERY) @Valid @RequestParam(value = "applicationDate", required = true) String applicationDate
    );

}